% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmera_functions.R
\name{calculate_derivative}
\alias{calculate_derivative}
\title{Calculate derivatives}
\usage{
calculate_derivative(link, pred)
}
\arguments{
\item{link}{Link function. Extracted from object}

\item{pred}{Linear predictor X\%*\%betas}
}
\value{
A vector of derivatives
}
\description{
Calculate derivatives
}
