% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmera_functions.R
\name{calculate_z_statistics}
\alias{calculate_z_statistics}
\title{Calculate z-statistics}
\usage{
calculate_z_statistics(mean, se)
}
\arguments{
\item{mean}{Mean value}

\item{se}{Standard Error}
}
\value{
Vector of mean, SE, z statistic, 95\% CI, and p-value
}
\description{
Calculate z-statistics
}
\keyword{internal}
