% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmera_functions.R
\name{contrast_mean}
\alias{contrast_mean}
\title{Calculate a mean for a single contrast}
\usage{
contrast_mean(
  contrast,
  X = X,
  betas = betas,
  var_betas = var_betas,
  link = link
)
}
\arguments{
\item{contrast}{A vector specifying fixed values of covariates and variables
to be averaged over when (NA) calculating a margin}

\item{X}{The matrix of covariates used in the model}

\item{betas}{Marginalized regression coefficients}

\item{var_betas}{Variance covariance matrix of betas}

\item{link}{Link function. Extracted from object}
}
\value{
A list containing the mean, its SE, and a vector of Jacobians
}
\description{
Calculate a mean for a single contrast
}
