% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_means_function.R
\name{gee.means}
\alias{gee.means}
\title{Calculate margins and their differences from longitudinal mixed-effects regression models}
\usage{
gee.means(
  object,
  tx.contrast0 = NULL,
  tx.contrast1 = NULL,
  ctrl.contrast0 = NULL,
  ctrl.contrast1 = NULL
)
}
\arguments{
\item{object}{An object of type "MixMod" that is obtained from the
GLMMadaptive package}

\item{tx.contrast0}{A vector specifying fixed values of covariates and variables
to be averaged over when (NA) calculating a margin for treatment group at baseline.
Must be the same length as the coefficient vector from the object model.}

\item{tx.contrast1}{A vector to calculate the margin for the treatment group at follow-up.}

\item{ctrl.contrast0}{A vector to calculate the margin for the control group at baseline.}

\item{ctrl.contrast1}{A vector to calculate the margin for the control group at follow-up.}
}
\value{
A data frame of margins, their standard errors, z-statistics, and p-values.
If a contrast is specified for both baseline and follow-up, their difference is
also calculated. If all 4 contrasts are specified, the difference in differences
is also calculated.
}
\description{
Calculate margins and their differences from longitudinal mixed-effects regression models
}
