% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmera_functions.R
\name{group_means}
\alias{group_means}
\title{Calculate means and their difference for two contrasts}
\usage{
group_means(
  contrast0 = NULL,
  contrast1 = NULL,
  X = X,
  betas = betas,
  var_betas = var_betas,
  link = link
)
}
\arguments{
\item{contrast0}{A vector specifying fixed values of covariates and variables
to be averaged over when (NA) calculating a margin for treatment group at baseline}

\item{contrast1}{A contrast for a second time point}

\item{X}{The matrix of covariates used in the model}

\item{betas}{Marginalized regression coefficients}

\item{var_betas}{Variance covariance matrix of betas}

\item{link}{Link function. Extracted from object}
}
\value{
A list of means, SEs, Jacobians, for the two contrasts and their difference
}
\description{
Calculate means and their difference for two contrasts
}
