% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gruder}
\alias{gruder}
\title{Gruder smoking cessation data}
\format{
A data frame with 1744 rows and 8 columns:
\describe{
\item{id}{ID variable}
\item{quit}{0 if smoking, 1 if abstinent}
\item{time}{0, 1, 2, 4 for post-intervention, 6-, 12-, and 24-month follow-ups}
\item{group}{Group. 0 for control, 1 for intervention}
\item{grouptime}{grp x time}
\item{race}{White race. 0 for non-white, 1 for white}
\item{tv}{0 for not watching tv, 1 for watching}
\item{manual}{manual = 0 for not reading manual, 1 for reading manual}
...
}
}
\usage{
gruder
}
\description{
Smoking cessation data from Gruder et al. (1993)
}
\references{
Gruder CL, Mermelstein RJ, Kirkendol S, Hedeker D, Wong SC, Schreckengost J, Warnecke RB, Burzette R, Miller TQ.
Effects of social support and relapse prevention training as adjuncts to a televised smoking-cessation intervention. Journal
of Consulting and Clinical Psychology 1993; 61:113–120.
}
\keyword{datasets}
