% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_means_function.R
\name{long_means}
\alias{long_means}
\title{Calculate margins and their differences from longitudinal mixed-effects or GEE
regression models}
\usage{
long_means(
  object,
  tx.contrast0 = NULL,
  tx.contrast1 = NULL,
  ctrl.contrast0 = NULL,
  ctrl.contrast1 = NULL
)
}
\arguments{
\item{object}{An object of type "MixMod" that is obtained from the
GLMMadaptive package or an object of type "geeglm" that is obtained
from the gee package.}

\item{tx.contrast0}{A vector specifying fixed values of covariates and variables
to be averaged over when (NA) calculating a margin for treatment group at baseline.
Must be the same length as the coefficient vector from the object model.}

\item{tx.contrast1}{A vector to calculate the margin for the treatment group at follow-up.}

\item{ctrl.contrast0}{A vector to calculate the margin for the control group at baseline.}

\item{ctrl.contrast1}{A vector to calculate the margin for the control group at follow-up.}
}
\value{
A data frame of margins, their standard errors, z-statistics, and p-values.
If a contrast is specified for both baseline and follow-up, their difference is
also calculated. If all 4 contrasts are specified, the difference in differences
is also calculated.
}
\description{
Calculate margins and their differences from longitudinal mixed-effects or GEE
regression models
}
\examples{
# Fit a mixed-effects model using the GLMMadaptive backage
library(GLMMadaptive)
fm1 <- mixed_model(fixed = quit ~ group + time + grouptime + race + tv + manual,
random = ~ time | id, data = gruder,
family = binomial())
# Specify contrasts for means
tx.contrast0=c(1, 1, 0, 0, NA, NA, NA) # tx time
tx.contrast1=c(1, 1, 4, 4, NA, NA, NA) # tx time 4
ctrl.contrast0=c(1, 0, 0, 0, NA, NA, NA) #ctrl time 0
ctrl.contrast1=c(1, 0, 4, 0, NA, NA, NA) #ctrl time 4
# Enter model object and contrasts into the long_means function
output <- long_means(fm1, tx.contrast0=tx.contrast0, tx.contrast1=tx.contrast1,
                    ctrl.contrast0=ctrl.contrast0, ctrl.contrast1=ctrl.contrast1)

# Using a GEE model with the same contrasts
if (requireNamespace("geepack", quietly = TRUE)) {
library(geepack)
mod1 <- geeglm(quit ~ group + time + grouptime + race + tv + manual, id=id,
data=gruder, corstr="unstructured", family=binomial())

output <- long_means(fm1, tx.contrast0=tx.contrast0, tx.contrast1=tx.contrast1,
                    ctrl.contrast0=ctrl.contrast0, ctrl.contrast1=ctrl.contrast1)
output
}
}
