% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_means_function.R
\name{mix_means}
\alias{mix_means}
\title{Calculate margins and their differences from longitudinal mixed-effects regression models}
\usage{
mix_means(
  object,
  marginals,
  tx.contrast0 = NULL,
  tx.contrast1 = NULL,
  ctrl.contrast0 = NULL,
  ctrl.contrast1 = NULL
)
}
\arguments{
\item{object}{An object of type "MixMod" that is obtained from the
GLMMadaptive package}

\item{marginals}{An optional object of type "m_coefs" that is extracted
from a "MixMod" object using the marginal_coefs() function from
the GLMMadaptive package. See details below.}

\item{tx.contrast0}{A vector specifying fixed values of covariates and variables
to be averaged over when (NA) calculating a margin for treatment group at baseline.
Must be the same length as the coefficient vector from the object model.}

\item{tx.contrast1}{A vector to calculate the margin for the treatment group at follow-up.}

\item{ctrl.contrast0}{A vector to calculate the margin for the control group at baseline.}

\item{ctrl.contrast1}{A vector to calculate the margin for the control group at follow-up.}
}
\value{
A data frame of margins, their standard errors, z-statistics, and p-values.
If a contrast is specified for both baseline and follow-up, their difference is
also calculated. If all 4 contrasts are specified, the difference in differences
is also calculated.
}
\description{
Calculate margins and their differences from longitudinal mixed-effects regression models
}
\details{
The optional marginals argument is designed when the user is going to
make repeated calls to the mix_means() function using the same object.
By including the marginalized coefficient vector and its variance covariance
matrix as an argument, it will not have to be extracted from the object
every time the function is called, saving a considerable amount of time.
}
